pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Project #2 - Difference Tool"
authors = ["Alina Makarovskaya <makarovskaya1307@gmail.com>"]
repository = "https://github.com/makaralina/python-project-50.git"
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
gendiff= "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
